move Nodes directory to subcomponent of Tree component
