+++ src/components/Tree/Nodes/BaseNode/BaseNode.tsx
@@ -1,10 +1,9 @@
+import styles from 'components/Tree/Nodes/BaseNode/baseNode.module.css';
+import { DragHandle } from 'components/Tree/Nodes/BaseNode/DragHandle/DragHandle';
 import { useTreeDirection } from 'hooks';
 import { ReactNode, useEffect } from 'react';
 import { Handle, NodeProps, Position, useUpdateNodeInternals } from 'reactflow';
 import { DecisionStatus } from 'store/DecisionSlice/decisionSlice';
-
-import styles from './baseNode.module.css';
-import { DragHandle } from './DragHandle/DragHandle';
 
 interface BaseNodeProps extends Omit<NodeProps, 'data'> {
   children: ReactNode;
