+++ src/components/Tree/Nodes/BoolNode/bool.module.css
