+++ src/components/Tree/Nodes/DefaultNode/DefaultNode.spec.tsx
@@ -1,8 +1,8 @@
 import '@testing-library/jest-dom';
 import { cleanup, render, screen } from '@testing-library/react';
+import { DefaultNode } from 'components/Tree/Nodes/DefaultNode/DefaultNode';
 import { ReactFlowProvider } from 'reactflow';
 import { afterEach, describe, expect, test } from 'vitest';
-import { DefaultNode } from './DefaultNode';
 
 afterEach(() => cleanup());
