+++ src/components/Tree/index.ts
@@ -1,0 +1,4 @@
+export { Tree } from './Tree';
+export { BoolNode } from './Nodes/BoolNode/BoolNode';
+export type { BoolNodeData } from './Nodes/BoolNode/BoolNode';
+export { DefaultNode } from './Nodes/DefaultNode/DefaultNode';
