+++ src/hooks/useFetchConfig/useFetchConfig.tsx
@@ -1,4 +1,4 @@
-import { BoolNodeData } from 'components/Nodes/BoolNode/BoolNode';
+import { BoolNodeData } from 'components/Tree';
 import { useEffect, useState } from 'react';
 import { PositionUnawareDecisionTree, TreeNode } from 'store';
 import { BooleanNodeData, NodeData } from 'store/DecisionSlice/decisionSlice';
